Remove use of std namespace and create debug print

+++ src/include/utils.h
@@ -1,23 +1,18 @@
 #pragma once
 #include "./common.h"
-
+#include <windows.h> // Windows
 
 // #ifdef _WIN32
 //   #include <windows.h> // Windows
-//   void sleep(int seconds) {
-//     Sleep(seconds * 1e3); // milliseconds
-//   }
+//   Sleep(seconds * 1e3); // milliseconds
 // #else
 //   #include <unistd.h> // Linux
-//   void sleep(int seconds) {
-//     usleep(seconds * 1e6); // microseconds
-//   }
+//   usleep(seconds * 1e6); // microseconds
 // #endif
-
 
 class Utils
 {
 public:
-  static void sleep(int time);
+  static void sleep(double time);
   static double randomDouble(double min, double max);
 };
